refactor(web): centralize sortByName and filterBySearch in shared list utils
- Add lib/list/shared-filters with generic sortByName/filterBySearch
- Update products/sauces/recipes filter modules to use shared implementations
- Remove unused local helpers and Fuse wiring where appropriate
